wip - fixing nexus_create_destroy test failure
